Mech_eap: only output debugging when GSSEAP_TRACE is set
Previously we only output informational messages from libeap.
However, we also output them all the time to stdout, which is bad
because it disrupts the output of the program.  Now, only output
debugging when the GSSEAP_TRACE environment variable is set.  In that
case output all the libeap debugging, not just informational messages.

In addition, use secure_getenv if it is available to avoid introducing yet more issues if run in a raised privilege situation.