Fix for https://bugs.launchpad.net/moonshot/+bug/1705714 (Infinite loop in tids exit handlers)
Refactored the C++ static object whose destructor was unloading the libraries.
The shibresolver library is now unloaded by a "smaller" static object that is
created after shibresolver is loaded. Creating that object late means that it
will be destroyed earlier, before the static objects in the log4shib library
are destroyed.